requirements.txt
anthropic>=0.39.0
python-dotenv>=1.0.0
tqdm>=4.66.0
streamlit>=1.28.0
pandas>=2.0.0
kagglehub